Rendering(Ranking) adjust data rendering to new response format
