--- client/src/components/Dashboard/CommunitySection/api/fetchRanking.api.ts
@@ -1,4 +1,0 @@
-export const fetchRanking = async () => {
-  const response = await fetch('/api/ranking');
-  return response.json();
-};
